+++ src/Compute/Compute.Test/ScenarioTests/DiskRPTests.ps1
@@ -1864,3 +1864,90 @@ function Test-DiskGrantAccessGetSASWithTL
 		Clean-ResourceGroup $rgname;
 	}
 }
+
+<#
+.SYNOPSIS
+Test confidential vm securityMetadataUri during confidential VM OS disk creation from an unmanaged storage account.
+#>
+function Test-ConfVMImportSecure
+{
+    # Setup
+    $rgname = Get-ComputeTestResourceName;
+    $loc = "eastus2euap";
+
+    try
+    {
+        New-AzResourceGroup -Name $rgname -Location $loc -Force;
+    
+        $rgname="haagha-test-gdskaccess"; 
+
+        $accessuri="https://haaghagsdkaccess.blob.core.windows.net/disks/access.vhd"; 
+        $securityuri="https://haaghagsdkaccess.blob.core.windows.net/disks/securitydata.vhd"; 
+        $securitymetadatauri="https://haaghagsdkaccess.blob.core.windows.net/disks/securitymetadata.vhd";
+        $storageacctid = "/subscriptions/88fd8cb2-8248-499e-9a2d-4929a4b0133c/resourceGroups/haagha-test-gdskaccess/providers/Microsoft.Storage/storageAccounts/haaghagsdkaccess"
+        $securityTypeDSP = "ConfidentialVM_DiskEncryptedWithPlatformKey";
+        $diskName = "testDiskconfv2"
+
+        $diskConfig = New-AzDiskConfig -Location $loc -CreateOption ImportSecure -SourceUri $accessuri -SecurityDataUri $securityuri -SecurityMetadataUri $securitymetadatauri -StorageAccountId $storageacctid -HyperVGeneration V2
+        Set-AzDiskSecurityProfile -Disk $diskConfig -SecurityType $securityTypeDSP 
+        
+        New-AzDisk -ResourceGroupName $rgname -DiskName $diskName -Disk $diskConfig -AsJob
+
+        $grantAccess = Grant-AzDiskAccess -ResourceGroupName $rgname -DiskName $diskName -Access 'Read' -DurationInSecond 60 -SecureVMGuestStateSAS;
+        Assert-NotNull $grantAccess.SecurityMetadataAccessSAS;
+        Assert-NotNull $grantAccess.SecurityDataAccessSAS;
+
+
+    }
+    finally
+    {
+        # Cleanup
+        Clean-ResourceGroup $rgname;
+    }
+}
+
+
+<#
+.SYNOPSIS
+Test confidential vm securityMetadataUri during confidential VM OS disk creation from an unmanaged storage account.
+#>
+function Test-DiskSnapshotInstantAccess
+{
+    # Setup
+    $rgname = Get-ComputeTestResourceName;
+    $location = "eastus2euap";
+
+    try
+    {
+        $diskName = "haagha-premiumv2test"
+        $snapshotName = "snapshotTest"
+
+        New-AzResourceGroup -Name $rgname -Location $location -Force;
+                
+        $diskConfig = New-AzDiskConfig `
+          -Location $location `
+          -DiskSizeGB 1024 `
+          -DiskIOPSReadWrite 10000 `
+          -DiskMBpsReadWrite 500 `
+          -AccountType PremiumV2_LRS `
+          -CreateOption Empty `
+          -Zone $zone
+
+        New-AzDisk -ResourceGroupName $rgname -DiskName $diskName -Disk $diskConfig
+
+        $disk = Get-AzDisk -ResourceGroupName $rgname -DiskName $diskName
+        $snapshotConfig = New-AzSnapshotConfig -SourceUri $disk.Id  -Location $location  -CreateOption Copy -InstantAccessDurationMinutes 300 -Incremental Premium_LRS
+
+        New-AzSnapshot -Snapshot $snapshotConfig -SnapshotName $snapshotName -ResourceGroupName $rgname
+
+        $snapshotGet = Get-AzSnapshot -ResourceGroupName $rgname -SnapshotName $snapshotName
+
+        Asset-NotNull = $snapshotGet.CreationData.InstantAccessDurationMinutes
+        Asset-NotNull = $snapshotGet.SnapshotAccessState
+    }
+    finally
+    {
+        # Cleanup
+        Clean-ResourceGroup $rgname;
+    }
+}