+++ src/Cdn/Cdn/ChangeLog.md
@@ -20,10 +20,12 @@
 
 ## Upcoming Release
 
-
 ## Version 5.1.0
 
 * Upgraded Api version to 2025-06-01
+## Version 5.0.1
+* Fixed content types comtent bug with null content
+* Added breaking change message for 2025-06-01
 
 ## Version 5.0.0
 * Upgraded Api version to 2025-04-15