+++ src/CosmosDB/CosmosDB/ChangeLog.md
@@ -20,6 +20,9 @@
 
 ## Upcoming Release
 
+## Version 1.18.2
+* Fixed issue where the restore command failed to detect the source account, even when it existed.
+
 ## Version 1.18.1
 * Fixed incorrect help documents for MongoDB cmdlets.