+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -19,6 +19,13 @@
 -->
 ## Upcoming Release
 
+## Version 7.9.0
+* Added Cross region restore support for new region - eastus3.
+* Added support for reregistration of a backup item to another vault.
+
+## Version 7.8.0
+* Added new optional parameter CVMOsDiskEncryptionSetId in Restore-AzRecoveryServicesBackupItem.
+
 ## Version 7.7.2
 * Added fix for resume protection with AzureFileshare.
 * Added fix for Get backup item for MSSQL soft deleted items.